refactor(web): extract runtime registry into pi-backend runtime
