+++ web/lib/rpc-manager.test.mjs
@@ -2,9 +2,12 @@ import assert from "node:assert/strict";
 import { readFile } from "node:fs/promises";
 import test from "node:test";
 
+const RUNTIME_MANAGER_URL = new URL("../packages/pi-backend/runtime-manager.ts", import.meta.url);
+const RUNTIME_URL = new URL("../packages/pi-backend/runtime.ts", import.meta.url);
+
 test("RPC session startup preloads extension-registered providers before restoring models", async () => {
-  const source = await readFile(new URL("./rpc-manager.ts", import.meta.url), "utf8");
-  const startupSource = source.slice(source.indexOf("export async function startRpcSession"));
+  const source = await readFile(RUNTIME_MANAGER_URL, "utf8");
+  const startupSource = source.slice(source.indexOf("export class RuntimeManager"));
 
   assert.match(startupSource, /createAgentSessionServices\(/);
   assert.match(startupSource, /createAgentSessionFromServices\(/);
@@ -12,12 +15,13 @@ test("RPC session startup preloads extension-registered providers before restori
 });
 
 test("RPC session startup resolves and passes the SDK-native enabled model scope", async () => {
-  const source = await readFile(new URL("./rpc-manager.ts", import.meta.url), "utf8");
-  const startupSource = source.slice(source.indexOf("export async function startRpcSession"));
+  const source = await readFile(RUNTIME_MANAGER_URL, "utf8");
+  const startupSource = source.slice(source.indexOf("export class RuntimeManager"));
   const resolveIndex = startupSource.indexOf("resolveVisibleModels(");
   const createIndex = startupSource.indexOf("createAgentSessionFromServices(");
 
   assert.ok(resolveIndex >= 0);
+  assert.ok(createIndex >= 0);
   assert.ok(createIndex > resolveIndex);
   assert.match(startupSource, /selectInitialModelScope\(/);
   assert.match(startupSource, /scopedModels: initial\.scopedModels/);
@@ -26,8 +30,8 @@ test("RPC session startup resolves and passes the SDK-native enabled model scope
 });
 
 test("RPC session startup treats only sessions with messages as continuing", async () => {
-  const source = await readFile(new URL("./rpc-manager.ts", import.meta.url), "utf8");
-  const startupSource = source.slice(source.indexOf("export async function startRpcSession"));
+  const source = await readFile(RUNTIME_MANAGER_URL, "utf8");
+  const startupSource = source.slice(source.indexOf("export class RuntimeManager"));
 
   assert.match(
     startupSource,
@@ -39,8 +43,8 @@ test("RPC session startup treats only sessions with messages as continuing", asy
 });
 
 test("RPC session startup opens an existing session file only once and trusts its cwd", async () => {
-  const source = await readFile(new URL("./rpc-manager.ts", import.meta.url), "utf8");
-  const startupSource = source.slice(source.indexOf("export async function startRpcSession"));
+  const source = await readFile(RUNTIME_MANAGER_URL, "utf8");
+  const startupSource = source.slice(source.indexOf("export class RuntimeManager"));
   const routeSource = await readFile(new URL("../app/api/agent/[id]/route.ts", import.meta.url), "utf8");
   const eventRouteSource = await readFile(new URL("../app/api/agent/[id]/events/route.ts", import.meta.url), "utf8");
   const autoNameRouteSource = await readFile(new URL("../app/api/sessions/[id]/auto-name/route.ts", import.meta.url), "utf8");
@@ -55,15 +59,15 @@ test("RPC session startup opens an existing session file only once and trusts it
 });
 
 test("RPC wrapper avoids per-chunk idle and running-state maintenance", async () => {
-  const wrapperSource = await readFile(new URL("../packages/pi-backend/runtime.ts", import.meta.url), "utf8");
-  const source = await readFile(new URL("./rpc-manager.ts", import.meta.url), "utf8");
+  const wrapperSource = await readFile(RUNTIME_URL, "utf8");
+  const managerSource = await readFile(RUNTIME_MANAGER_URL, "utf8");
   const startSource = wrapperSource.slice(
     wrapperSource.indexOf("  start(): void"),
     wrapperSource.indexOf("  setForceEmptySystemPrompt"),
   );
-  const notifySource = source.slice(
-    source.indexOf("export function notifyRunningChange"),
-    source.indexOf("export async function startRpcSession"),
+  const notifySource = managerSource.slice(
+    managerSource.indexOf("notifyRunningChange(): void {"),
+    managerSource.indexOf("async startSession("),
   );
 
   assert.match(startSource, /IDLE_RESET_EVENT_TYPES\.has\(event\.type\)/);
@@ -74,7 +78,7 @@ test("RPC wrapper avoids per-chunk idle and running-state maintenance", async ()
 });
 
 test("normal session teardown paths use graceful extension shutdown", async () => {
-  const source = await readFile(new URL("../packages/pi-backend/runtime.ts", import.meta.url), "utf8");
+  const source = await readFile(RUNTIME_URL, "utf8");
   const deleteRouteSource = await readFile(new URL("../app/api/sessions/[id]/route.ts", import.meta.url), "utf8");
   const trustRouteSource = await readFile(new URL("../app/api/project-trust/route.ts", import.meta.url), "utf8");
   const idleSource = source.slice(
@@ -116,15 +120,15 @@ test("prompt routes mark only preflight failures as rejected", async () => {
 });
 
 test("RPC session startup persists explicit preferences without replaying setters", async () => {
-  const source = await readFile(new URL("./rpc-manager.ts", import.meta.url), "utf8");
-  const startupSource = source.slice(source.indexOf("export async function startRpcSession"));
+  const source = await readFile(RUNTIME_MANAGER_URL, "utf8");
+  const startupSource = source.slice(source.indexOf("export class RuntimeManager"));
 
   assert.match(startupSource, /persistExplicitStartupPreferences\(/);
   assert.match(startupSource, /modelDefaultChanged\) invalidateModelsCache\(\)/);
 });
 
 test("custom extension UI receives the fixed headless terminal facade", async () => {
-  const source = await readFile(new URL("../packages/pi-backend/runtime.ts", import.meta.url), "utf8");
+  const source = await readFile(RUNTIME_URL, "utf8");
   const customUiSource = source.slice(
     source.indexOf("private requestExtensionCustomUi"),
     source.indexOf("private requestExtensionUi"),
@@ -135,7 +139,7 @@ test("custom extension UI receives the fixed headless terminal facade", async ()
 });
 
 test("reloading a session invalidates the models cache", async () => {
-  const source = await readFile(new URL("../packages/pi-backend/runtime.ts", import.meta.url), "utf8");
+  const source = await readFile(RUNTIME_URL, "utf8");
   const reloadSource = source.slice(
     source.indexOf('case "reload"'),
     source.indexOf('case "abort_compaction"'),
@@ -143,4 +147,28 @@ test("reloading a session invalidates the models cache", async () => {
 
   assert.match(reloadSource, /await this\.inner\.reload\(\)/);
   assert.match(reloadSource, /this\.applyForcedEmptySystemPrompt\(\);\s*invalidateModelsCache\(\)/);
+});
+
+test("rpc-manager seam re-exports the frozen surface and delegates to the manager", async () => {
+  const source = await readFile(new URL("./rpc-manager.ts", import.meta.url), "utf8");
+
+  // Re-exported types/values (unchanged public surface).
+  assert.match(source, /export type \{ AgentEvent, AgentRuntimeHooks \} from/);
+  assert.match(source, /export \{ AgentSessionWrapper \} from/);
+  assert.match(source, /export type \{ RpcSessionStartOptions \} from/);
+
+  // Delegation seam: no registry/local helpers, everything flows through the
+  // manager singleton so the globalThis contract lives only in runtime-state.ts.
+  assert.match(source, /export function getRpcSession\(/);
+  assert.match(source, /export function getRpcSessionInfos\(/);
+  assert.match(source, /export function hasBusyRpcSessionForCwd\(/);
+  assert.match(source, /export async function destroyRpcSessionsForCwd\(/);
+  assert.match(source, /export function getRunningRpcSessionIds\(/);
+  assert.match(source, /export function subscribeRunningSessions\(/);
+  assert.match(source, /export function notifyRunningChange\(/);
+  assert.match(source, /export async function startRpcSession\(/);
+  assert.doesNotMatch(source, /getRegistry\(\)|getLocks\(\)|getStartingSessionCwds\(\)/);
+  assert.match(source, /getRuntimeManager\(\)\.getSession\(/);
+  assert.match(source, /getRuntimeManager\(\)\.getSessionInfos\(\)/);
+  assert.match(source, /getRuntimeManager\(\)\.startSession\(/);
 });
